+++ src/chains/WorldChain.sol
@@ -18,6 +18,8 @@ import {
     uniswapV3ForkId,
     IUniswapV3Callback
 } from "../core/univ3forks/UniswapV3.sol";
+import {dackieSwapV3WorldChainFactory, dackieSwapV3ForkId} from "../core/univ3forks/DackieSwapV3.sol";
+import {pancakeSwapV3InitHash, IPancakeSwapV3Callback} from "../core/univ3forks/PancakeSwapV3.sol";
 
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../SettlerAbstract.sol";
@@ -54,6 +56,10 @@ abstract contract WorldChainMixin is FreeMemory, SettlerBase {
             factory = uniswapV3WorldChainFactory;
             initHash = uniswapV3InitHash;
             callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == dackieSwapV3ForkId) {
+            factory = dackieSwapV3WorldChainFactory;
+            initHash = pancakeSwapV3InitHash;
+            callbackSelector = uint32(IPancakeSwapV3Callback.pancakeV3SwapCallback.selector);
         } else {
             revert UnknownForkId(forkId);
         }